requirements.txt
asgiref==3.4.1
cycler==0.11.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
opencv-python==4.5.4.58
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
